fix : 리뷰반영 - Result 클래스 삭제 후 BlackJackGame 클래스에서 게임의 결과도 컨트롤 하도록 수정
