+++ src/main/kotlin/blackjack/domain/GameResult.kt
@@ -1,0 +1,5 @@
+package blackjack.domain
+
+enum class GameResult() {
+    WIN, LOSE, DRAW
+}
